Rearrange all the files in respetive folders
