+++ Java/Encrypt Logic.java
