+++ Javascript/numberOfDivisors.js
