+++ Python/Flappy game inspired by Flappy Bird.py
